Atomic design toegepast
