+++ src/lib/Molecul/SDG.svelte
@@ -22,7 +22,7 @@
 	  <img
 		class="susDevGoal"
 		src={data.wish.image.image.url}
-		alt={data.wish.label}
+		alt=""
 		width="50"
 		height="50"
 		loading="lazy"
