+++ src/routes/+error.svelte
@@ -1,6 +1,5 @@
 <script>
   import { goto } from '$app/navigation';
-  import { page } from '$app/stores';
 </script>
 
 <div class="error-container">
