+++ src/routes/+page.server.js
@@ -21,12 +21,10 @@ export async function load() {
         url
         image {
           url
-          label
-        }
-        label
         }
       }
     }
+  }
   `
 
   const request = await hygraph.request(query)
