+++ src/routes/+page.svelte
@@ -1,11 +1,11 @@
 <script>
     export let data;
     import { onMount } from "svelte";
-    import Loader from '$lib/components/Loader.svelte';
-    import Navbar from '$lib/components/Navbar.svelte';
-    import Hero from '$lib/components/Hero.svelte';
-    import Carousel from '$lib/components/Carousel.svelte';
-    import Footer from '$lib/components/Footer.svelte';
+    import Loader from '$lib/Organism/Loader.svelte';
+    import Navbar from '$lib/Organism/Navbar.svelte';
+    import Hero from '$lib/Organism/Hero.svelte';
+    import Carousel from '$lib/Organism/Carousel.svelte';
+    import Footer from '$lib/Organism/Footer.svelte';
     
     onMount(() => {});
 </script>
